requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
aiohttp==3.9.5
    # via python-binance
aiosignal==1.3.1
    # via aiohttp
altair==5.3.0
    # via streamlit
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.8.2
    # via streamlit
cachetools==5.3.3
    # via streamlit
certifi==2024.6.2
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via streamlit
dateparser==1.2.0
    # via python-binance
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via streamlit
idna==3.7
    # via
    #   requests
    #   yarl
jinja2==3.1.4
    # via
    #   altair
    #   pydeck
jsonschema==4.22.0
    # via altair
jsonschema-specifications==2023.12.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
numpy==2.0.0
    # via
    #   altair
    #   pandas
    #   pyarrow
    #   pydeck
    #   streamlit
packaging==24.1
    # via
    #   altair
    #   plotly
    #   streamlit
pandas==2.2.2
    # via
    #   altair
    #   streamlit
pillow==10.3.0
    # via streamlit
plotly==5.22.0
    # via -r requirements.in
protobuf==5.27.2
    # via streamlit
pyarrow==16.1.0
    # via streamlit
pycryptodome==3.20.0
    # via python-binance
pydeck==0.9.1
    # via streamlit
pygments==2.18.0
    # via rich
python-binance==1.0.19
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   dateparser
    #   pandas
pytz==2024.1
    # via
    #   dateparser
    #   pandas
referencing==0.35.1
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.5.15
    # via dateparser
requests==2.32.3
    # via
    #   python-binance
    #   streamlit
rich==13.7.1
    # via streamlit
rpds-py==0.18.1
    # via
    #   jsonschema
    #   referencing
six==1.16.0
    # via
    #   python-binance
    #   python-dateutil
smmap==5.0.1
    # via gitdb
streamlit==1.36.0
    # via -r requirements.in
tenacity==8.4.2
    # via
    #   plotly
    #   streamlit
toml==0.10.2
    # via streamlit
toolz==0.12.1
    # via
    #   -r requirements.in
    #   altair
tornado==6.4.1
    # via streamlit
typing-extensions==4.12.2
    # via
    #   altair
    #   streamlit
tzdata==2024.1
    # via pandas
tzlocal==5.2
    # via dateparser
ujson==5.10.0
    # via python-binance
urllib3==2.2.2
    # via requests
websockets==12.0
    # via python-binance
yarl==1.9.4
    # via aiohttp
